Trying fetch data to view
