+++ core/UserModel.php
@@ -15,5 +15,27 @@ abstract class UserModel extends DbModel
 {
     abstract public function getDisplayName(): string;
 
+    public static function findAll()
+    {
+        $tableName = static::tableName();
+        // $attributes = array_keys($where);
+
+        $statement = self::prepare("SELECT id, firstname, lastname, created_at FROM $tableName");
+
+        $statement->execute();
+        return $statement->fetchAll(\PDO::FETCH_OBJ);
+    }
+
+    public function assignDefaultRole()
+    {
+        $tableName = static::tableName();
+
+        $statement = self::prepare("SELECT * FROM $tableName;");
+
+        $statement->execute();
+
+        return $statement->fetchObject(static::class);
+    }
+
 }
 ?>
